+++ src/utils/feedContent.ts
@@ -1,6 +1,6 @@
 import { experimental_AstroContainer as AstroContainer } from 'astro/container';
 import { loadRenderers } from 'astro:container';
-import { getContainerRenderer as getMDXRenderer } from '@astrojs/mdx';
+import { getContainerRenderer as getMDXRenderer } from '@astrojs/mdx/container-renderer';
 import { render, type CollectionEntry } from 'astro:content';
 import { SEO_CONFIG } from '../consts';
 import { generateImageUrl } from './seo';